+++ src/main/preload.ts
@@ -1,8 +1,12 @@
 import { contextBridge, ipcRenderer } from 'electron';
 
 contextBridge.exposeInMainWorld('electronAPI', {
-    // Add IPC methods here
     ping: () => ipcRenderer.invoke('ping'),
     selectFile: () => ipcRenderer.invoke('dialog:openFile'),
     selectDirectory: () => ipcRenderer.invoke('dialog:openDirectory'),
+    getBackendPort: () => ipcRenderer.invoke('get-backend-port'),
+    getLogPath: () => ipcRenderer.invoke('get-log-path'),
+    installUpdate: () => ipcRenderer.invoke('install-update'),
+    onUpdateDownloaded: (callback: (version: string) => void) =>
+        ipcRenderer.on('update-downloaded', (_event, version) => callback(version)),
 });